Add chroma
