+++ configs/model/components/chroma.yaml
@@ -1,4 +1,4 @@
-_target_: src.models.components.Searchers.FAISSSearcher
+_target_: src.models.components.Searchers.ChromaSearcher
 chunk_size: 10000
 chunk_overlap: 0
 emb_model: "text-embedding-ada-002"
