+++ configs/model/default.yaml
@@ -1,3 +1,3 @@
 # seacherクラスのデフォルト設定を記述します
 defaults:
-  - components: FAISS.yaml
+  - components: chroma.yaml
